adds statefulset reconciler
