+++ pkg/controllers/dynakube/extension/eec/reconciler.go
@@ -1,0 +1,31 @@
+package eec
+
+import (
+	"github.com/Dynatrace/dynatrace-operator/pkg/api/v1beta3/dynakube"
+	"github.com/Dynatrace/dynatrace-operator/pkg/controllers"
+	"golang.org/x/net/context"
+	"sigs.k8s.io/controller-runtime/pkg/client"
+)
+
+type reconciler struct {
+	client    client.Client
+	apiReader client.Reader
+
+	dk *dynakube.DynaKube
+}
+
+type ReconcilerBuilder func(clt client.Client, apiReader client.Reader, dk *dynakube.DynaKube) controllers.Reconciler
+
+var _ ReconcilerBuilder = NewReconciler
+
+func NewReconciler(clt client.Client, apiReader client.Reader, dk *dynakube.DynaKube) controllers.Reconciler {
+	return &reconciler{
+		client:    clt,
+		apiReader: apiReader,
+		dk:        dk,
+	}
+}
+
+func (r *reconciler) Reconcile(ctx context.Context) error {
+	return r.reconcileStatefulset(ctx)
+}
